refactor buildah types and structs from heavy deps
When other pieces of code, like the go bindings in libpod, import buildah structs and types, they do not want to also have the burden of importing containers image and containers storage.  To avoid this, we put these structs into builld/define/run|build.go.
